Test console output run (#33)
* Works to do the launch command in a console window for windows. To tweak for the other OS's

Works to do the launch command in a console window for windows. To tweak for the other OS's

Works to do the launch command in a console window for windows. To tweak for the other OS's

Works to do the launch command in a console window for windows. To tweak for the other OS's

Works to do the launch command in a console window for windows. To tweak for the other OS's

The delete button now disables the start stop buttons etc...

Made the other operations also change as required

* Messed up the linux launch code